Update the code to fix build issues
Update the code to fix build issues
