+++ src/app/layout.tsx
@@ -1,8 +1,10 @@
 import { Analytics } from '@vercel/analytics/react';
+import dynamic from 'next/dynamic';
 import { Inter } from 'next/font/google';
 
 import { PHProvider } from '@/lib/PHProvider';
 import Provider from '@/lib/Provider';
+
 import './globals.css';
 
 export const viewport = {
@@ -11,6 +13,10 @@ export const viewport = {
   maximumScale: 1,
   userScalable: false,
 };
+
+const PostHogPageView = dynamic(() => import('@/lib/PostHogPageView'), {
+  ssr: false,
+});
 
 const inter = Inter({ subsets: ['latin'], adjustFontFallback: false });
 
@@ -24,6 +30,7 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <PHProvider>
         <body className={`${inter.className} bg-muted text-foreground`}>
+          <PostHogPageView />
           <Provider>{children}</Provider>
           <Analytics />
         </body>
